# faturamento-api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API RESTful para emissao, gerenciamento de Notas Fiscais e integracao com SQL Server.",
        "title": "API de Faturamento (Microsservico Go)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte Korp",
            "url": "https://github.com/pheiterer/Korp_Teste_Paulo"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/api/v1/notas-fiscais": {
            "get": {
                "description": "Retorna a lista de notas fiscais paginada com seus respectivos itens (query params: page, limit, status).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notas Fiscais"
                ],
                "summary": "Listar Notas Fiscais com Paginação e Filtros",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número da página (padrão: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade de itens por página (padrão: 10, máx: 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filtro por status (Aberta, EmProcessamento, Fechada, Cancelada)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova Nota Fiscal no banco de dados com status 'Aberta'.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notas Fiscais"
                ],
                "summary": "Criar Nota Fiscal",
                "parameters": [
                    {
                        "description": "Payload de Criacao da Nota Fiscal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateNotaFiscalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notas-fiscais/{id}": {
            "get": {
                "description": "Retorna os detalhes de uma nota fiscal pelo seu ID ou UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notas Fiscais"
                ],
                "summary": "Buscar Nota Fiscal por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID ou UUID da Nota Fiscal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notas-fiscais/{id}/imprimir": {
            "post": {
                "description": "Altera o status da Nota Fiscal para 'Fechada' e dispara o evento assincrono NotaFiscalEmitidaEvent para o RabbitMQ.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notas Fiscais"
                ],
                "summary": "Imprimir Nota Fiscal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID ou UUID da Nota Fiscal",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Realiza a verificacao de saude da API de Faturamento e de suas dependencias de infraestrutura.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health Check da API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ComponentHealthResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ComponentHealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/handlers.ErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "handlers.ComponentHealth": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.ComponentHealthResponse": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/handlers.ComponentHealth"
                    }
                },
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateNotaFiscalItemRequest": {
            "type": "object",
            "required": [
                "quantidade"
            ],
            "properties": {
                "codigoProduto": {
                    "type": "string"
                },
                "codigo_produto": {
                    "type": "string"
                },
                "precoUnitario": {
                    "type": "number"
                },
                "preco_unitario": {
                    "type": "number"
                },
                "produtoId": {
                    "type": "integer"
                },
                "produto_id": {
                    "type": "integer"
                },
                "quantidade": {
                    "type": "integer"
                }
            }
        },
        "handlers.CreateNotaFiscalRequest": {
            "type": "object",
            "required": [
                "itens"
            ],
            "properties": {
                "itens": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/handlers.CreateNotaFiscalItemRequest"
                    }
                },
                "numero_sequencial": {
                    "type": "integer"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}